修改bug

+++ app/http/ajax.js
@@ -58,7 +58,9 @@ export function asyncAwaitCall(obj) {
     //TODO 登录不需要token,获取验证码和判断账号是否需要验证码都不需要token
     if (!token && obj.url.value.indexOf("login") === -1
                 && obj.url.value.indexOf("code") === -1
-                && obj.url.value.indexOf("region") === -1) {
+                && obj.url.value.indexOf("region") === -1
+                && obj.url.value.indexOf("register") === -1
+                && obj.url.value.indexOf("Password") === -1) {
         window.localStorage.clear();
         window.location = '/';
     }
